enhancement #485: moved endpoint for selective disciplines to the user controller
- moved selectDisciplines method to the user service
- moved private methods to the user service
- moved integration tests for selectDisciplines method to UserService.ispec.ts
- changed dependencies

closes #485
