+++ src/v2/api/services/UserService.ts
@@ -24,6 +24,18 @@ import { FileService } from '../../utils/files/FileService';
 import { DisciplineMapper } from '../../mappers/DisciplineMapper';
 import { RemainingSelectiveDTO } from '../dtos/RemainingSelectiveDTO';
 import { DateService } from '../../utils/date/DateService';
+import { AttachSelectiveDisciplinesDTO } from '../dtos/AttachSelectiveDisciplinesDTO';
+import { DbDiscipline } from '../../database/entities/DbDiscipline';
+import { NotBelongToGroupException } from '../../utils/exceptions/NotBelongToGroupException';
+import { ExcessiveSelectiveDisciplinesException } from '../../utils/exceptions/ExcessiveSelectiveDisciplinesException';
+import { checkIfArrayIsUnique } from '../../utils/ArrayUtil';
+import { AlreadySelectedException } from '../../utils/exceptions/AlreadySelectedException';
+
+type SortedDisciplines = {
+  year: number;
+  semester: number;
+  disciplines: string[];
+}
 
 @Injectable()
 export class UserService {
@@ -357,5 +369,105 @@ export class UserService {
       availableSelectiveAmount,
       remainingSelective,
     };
+  }
+
+  private checkIsEachDisciplineBelongUserGroup (
+    disciplines: DbDiscipline[],
+    groupId: string
+  ) {
+    disciplines.map((discipline) => {
+      if (discipline.groupId !== groupId) {
+        throw new NotBelongToGroupException();
+      }
+    });
+  }
+
+  private async getSelectiveDisciplines (userId: string) {
+    return this.disciplineRepository.findMany({
+      where: {
+        selectiveDisciplines: {
+          some: {
+            studentId: userId,
+          },
+        },
+      },
+    });
+  }
+
+  private async checkExcessiveSelectiveDisciplines (
+    pendingDisciplines: SortedDisciplines[],
+    selectedDisciplines: SortedDisciplines[],
+    groupId: string,
+  ) {
+    for (const { year, semester, disciplines } of pendingDisciplines) {
+      const pendingAmount = disciplines.length;
+      const selectedAmount = selectedDisciplines
+        .find((p) => p.year === year && p.semester === semester)
+        ?.disciplines.length ?? 0;
+      const { selectiveAmounts } = await this.groupRepository.find({
+        selectiveAmounts: {
+          some: {
+            groupId: groupId,
+            year: year,
+            semester: semester,
+          },
+        },
+      });
+      const { amount } = selectiveAmounts.find((s) => s.year === year && s.semester === semester);
+      if (pendingAmount + selectedAmount > amount) {
+        throw new ExcessiveSelectiveDisciplinesException();
+      }
+    }
+  }
+
+  private checkAlreadySelectedDisciplines (disciplineIds, selectedDisciplineIds) {
+    const isUnique = checkIfArrayIsUnique([...disciplineIds, ...selectedDisciplineIds]);
+    if (!isUnique) {
+      throw new AlreadySelectedException();
+    }
+  }
+
+  private async attachSelectiveDisciplines (
+    userId: string,
+    disciplineIds: string[]
+  ) {
+    await this.studentRepository.update({
+      where: {
+        userId: userId,
+      },
+      data: {
+        selectiveDisciplines: {
+          createMany: {
+            data: disciplineIds.map((disciplineId) => ({ disciplineId })),
+          },
+        },
+      },
+    });
+  }
+
+  async selectDisciplines (
+    userId: string,
+    body: AttachSelectiveDisciplinesDTO,
+  ) {
+    const disciplines = [];
+    for (const disciplineId of body.disciplines) {
+      disciplines.push(await this.disciplineRepository.findById(disciplineId));
+    }
+    const { id: groupId } = await this.groupRepository.find({
+      students: {
+        some: {
+          userId: userId,
+        },
+      },
+    });
+    this.checkIsEachDisciplineBelongUserGroup(disciplines, groupId);
+
+    const sortedDisciplines = this.disciplineMapper.getSortedDisciplinesByPeriod(disciplines);
+    const selectedDisciplines = await this.getSelectiveDisciplines(userId);
+    const sortedSelectedDisciplines = this.disciplineMapper.getSortedDisciplinesByPeriod(selectedDisciplines);
+
+    this.checkAlreadySelectedDisciplines(body.disciplines, selectedDisciplines.map((d) => d.id));
+    await this.checkExcessiveSelectiveDisciplines(sortedDisciplines, sortedSelectedDisciplines, groupId);
+    await this.attachSelectiveDisciplines(userId, body.disciplines);
   }
 }
